+++ atompack/__init__.py
@@ -1,3 +1,4 @@
 from atompack.atom import Atom
+from atompack.bond import Bond
 from atompack.crystal import Crystal, UnitCell
 from atompack.molecule import Molecule