+++ atompack/molecule.py
@@ -18,13 +18,29 @@ class Molecule(Topology):
 
     @classmethod
     def linear2(cls, atoms: List[Atom], bond_length: float) -> 'Molecule':
-        """Initializes a molecule with a pair/linear configuration."""
-        pass
+        """Initializes a molecule with an idealized pair/linear configuration.
+        
+        Note:
+            The `position` attribute of each atom is ignored and overwritten.
+        """
+        # atoms[0]._position = np.zeros(3)
+        # atoms[1]._position = np.array([bond_length, 0, 0])
+        # bonds = [(0, 1)]
+        # return cls(atoms, bonds)
 
     @classmethod
     def linear3(cls, core_atom: Atom, bonded_atoms: List[Atom], bond_lengths: List[float]) -> 'Molecule':
-        """Initializes a molecule with a triplet/linear configuration."""
-        pass
+        """Initializes a molecule with an idealized triplet/linear configuration.
+        
+        Note:
+            The `position` attribute of each atom is ignored and overwritten.
+        """
+        # bonded_atoms[0]._position = np.zeros(3)
+        # core_atom._position = np.array([bond_lengths[0], 0, 0])
+        # bonded_atoms[1]._position = np.array([bond_lengths[0] + bond_lengths[1], 0, 0])
+        # atoms = bonded_atoms + core_atom
+        # bonds = [(0, 2), (1, 2)]
+        # return cls(atoms, bonds)
 
     @classmethod
     def trigonal_planar(
@@ -33,8 +49,13 @@ class Molecule(Topology):
         bonded_atoms: List[Atom],
         bond_lengths: List[float],
     ) -> 'Molecule':
-        """Initializes a molecule with a trigonal planar configuration."""
+        """Initializes a molecule with an idealized trigonal planar configuration.
+        
+        Note:
+            The `position` attribute of each atom is ignored and overwritten.
+        """
         pass
+        
 
     @classmethod
     def bent(
@@ -44,7 +65,11 @@ class Molecule(Topology):
         bond_lengths: List[float],
         theta: float,
     ) -> 'Molecule':
-        """Initializes a molecule with a bent/angular configuration."""
+        """Initializes a molecule with an idealized bent/angular configuration.
+        
+        Note:
+            The `position` attribute of each atom is ignored and overwritten.
+        """
         pass
 
     @classmethod
@@ -52,9 +77,13 @@ class Molecule(Topology):
         cls,
         core_atom: Atom,
         bonded_atoms: List[Atom],
-        bond_lengths: List[Atom],
+        bond_lengths: List[float],
     ) -> 'Molecule':
-        """Initializes a molecule with a tetrahedral configuration."""
+        """Initializes a molecule with an idealized tetrahedral configuration.
+        
+        Note:
+            The `position` attribute of each atom is ignored and overwritten.
+        """
         pass
 
     @classmethod
@@ -65,7 +94,11 @@ class Molecule(Topology):
         bond_lengths: List[float],
         theta: float,
     ) -> 'Molecule':
-        """Initializes a molecule with a trigonal pyramidal configuration."""
+        """Initializes a molecule with an idealized trigonal pyramidal configuration.
+        
+        Note:
+            The `position` attribute of each atom is ignored and overwritten.
+        """
         pass
 
     @classmethod
@@ -77,7 +110,11 @@ class Molecule(Topology):
         axial_bond_lengths: List[float],
         lateral_bond_lengths: List[float],
     ) -> 'Molecule':
-        """Initializes a molecule with a trigonal bipyramidal configuration."""
+        """Initializes a molecule with an idealized trigonal bipyramidal configuration.
+        
+        Note:
+            The `position` attribute of each atom is ignored and overwritten.
+        """
         pass
 
     @classmethod
@@ -89,5 +126,9 @@ class Molecule(Topology):
         axial_bond_lengths: List[float],
         lateral_bond_lengths: List[float],
     ) -> 'Molecule':
-        """Initializes a molecule with an octahedral configuration."""
+        """Initializes a molecule with an idealized octahedral configuration.
+        
+        Note:
+            The `position` attribute of each atom is ignored and overwritten.
+        """
         pass