+++ atompack/crystal/__init__.py
@@ -1,9 +1,10 @@
 """
-Module `crystal` provides abstractions to construct crystals.
-At the lowest level, a generic lattice and atomic basis with no crystallographic constraints can be composed in a `UnitCell` object.
-At the mid-level, bravais lattices with specific symmetry constraints can be constructed with classes such as `Cubic`, `Orthorhombic`, or `Monoclinic`.
-At the highest level, users can create a number of common crystal structures such as `Fcc`, `Rutile`, or `Diamond`.
-For transformations like rotations, reorientations, and supercell duplication, the `Crystal` class wraps a `UnitCell` and modifies it with the desired transforms.
+Module `crystal` provides abstractions to construct crystals. At the lowest level, a generic lattice and atomic 
+basis with no crystallographic constraints can be composed in a `UnitCell` object. At the mid-level, bravais 
+lattices with specific symmetry constraints can be constructed with classes such as `Cubic`, `Orthorhombic`, or 
+`Monoclinic`. At the highest level, users can create a number of common crystal structures such as `Fcc`, `Rutile`, 
+or `Diamond`. For transformations like rotations, reorientations, and supercell duplication, the `Crystal` class 
+wraps a `UnitCell` object to produce a new topology with the transformations applied.
 """
 from atompack.crystal.crystal import Crystal
 from atompack.crystal.unit_cell import (CsCl, Cubic, Diamond, Fcc, Hcp, Hexagonal, Monoclinic, NaCl, Orthorhombic,