+++ CHANGELOG.md
@@ -9,8 +9,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* `Molecule` class.
+* `bond` module.
+* `molecule` module.
+* `topology` module.
+* `crystal.UnitCell` class.
 * Support for `pipenv` installation.
+* Additional benchmark tests.
+
+### Changed
+
+* `crystal.Crystal` now inherits from `topology.Topology`.
+* `_cell.cell_contains` implemented for paralellpipeds only.
+
+### Removed
+
+* `errors` module.
+* `structure` module.
 
 ## [0.2.1] - 2020-08-02