+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.3.0] - ???
+## [0.3.0] - 2020-09-08
 
 ### Added
 
@@ -19,10 +19,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 * `crystal.Crystal` now inherits from `topology.Topology`.
-* `_cell.cell_contains` implemented for paralellpipeds only.
 
 ### Removed
 
+* C extensions
 * `structure` module.
 
 ## [0.2.1] - 2020-08-02